Initial javascript setup.
