+++ .gitignore
@@ -1,6 +1,5 @@
-dist
-node_modules
+/dist/
+/node_modules/
+.vscode/
+
 secrets.js
-typings
-.idea
-.vscode
