--- src/_reference.ts
@@ -1,2 +1,0 @@
-/// <reference path="../typings/tsd.d.ts" />
-/// <reference path="../node_modules/screeps-typescript-declarations/dist/screeps.d.ts" />
