+++ src/gameManager.js
@@ -1,0 +1,3 @@
+export function loop() {
+    console.log('loop');
+}
