+++ src/main.js
@@ -1,0 +1,7 @@
+/* global module:true */
+
+import { loop as mainLoop } from './gameManager.js';
+
+module.exports.loop = function() {
+    mainLoop();
+};
