--- tsconfig.json
@@ -1,10 +1,0 @@
-{
-  "compilerOptions": {
-    "module": "commonjs",
-    "target": "es2015",
-    "noResolve": false,
-    "noImplicitAny": true,
-    "outDir": "./dist/"
-  }
-}
-
