--- tsd.json
@@ -1,12 +1,0 @@
-{
-  "version": "v4",
-  "repo": "borisyankov/DefinitelyTyped",
-  "ref": "master",
-  "path": "typings",
-  "bundle": "typings/tsd.d.ts",
-  "installed": {
-    "lodash/lodash.d.ts": {
-      "commit": "efa0c1196d7280640e624ac1e7fa604502e7bd63"
-    }
-  }
-}
